+++ lib/qonfig/data_set/class_builder.rb
@@ -23,18 +23,5 @@ module Qonfig::DataSet::ClassBuilder
         end
       end
     end
-
-    # @option base_klass [Class<Qonfig::DataSet>]
-    # @option child_klass [Class<Qonfig::DataSet>]
-    # @return [void]
-    #
-    # @api private
-    # @since 0.19.0
-    def inherit(base_klass:, child_klass:)
-      child_klass.definition_commands.concat(base_klass.definition_commands)
-      child_klass.instance_commands.concat(base_klass.instance_commands, &:inheritable?)
-      child_klass.predefined_validators.merge(base_klass.predefined_validators)
-      child_klass.validators.concat(base_klass.validators)
-    end
   end
 end