+++ lib/qonfig/dsl.rb
@@ -4,8 +4,10 @@
 # @since 0.1.0
 # @version 0.20.0
 module Qonfig::DSL # rubocop:disable Metrics/ModuleLength
+  require_relative 'dsl/inheritance'
+
   class << self
-    # @param child_klass [Qonfig::DataSet]
+    # @param child_klass [Class<Qonfig::DataSet>]
     # @return [void]
     #
     # @see Qonfig::DataSet::ClassBuilder
@@ -26,7 +28,7 @@ module Qonfig::DSL # rubocop:disable Metrics/ModuleLength
           child_klass.instance_variable_set(:@instance_commands, Qonfig::CommandSet.new)
           child_klass.instance_variable_set(:@predefined_validators, Qonfig::Validation::Collections::PredefinedRegistry.new)
           child_klass.instance_variable_set(:@validators, Qonfig::Validation::Collections::InstanceCollection.new)
-          Qonfig::DataSet::ClassBuilder.inherit(base_klass: self, child_klass: child_klass)
+          Qonfig::DSL::Inheritance.inherit(base: self, child: child_klass)
           super
         end
       end)
@@ -175,7 +177,7 @@ module Qonfig::DSL # rubocop:disable Metrics/ModuleLength
   # @since 0.2.0
   # @version 0.21.0
   def load_from_self(format: :dynamic)
-    caller_location = caller(1, 1).first
+    caller_location = ::Kernel.caller(1, 1).first
 
     definition_commands << Qonfig::Commands::Definition::LoadFromSelf.new(
       caller_location, format: format
@@ -252,7 +254,7 @@ module Qonfig::DSL # rubocop:disable Metrics/ModuleLength
   # @since 0.14.0
   # @version 0.21.0
   def expose_self(env:, format: :dynamic)
-    caller_location = caller(1, 1).first
+    caller_location = ::Kernel.caller(1, 1).first
 
     definition_commands << Qonfig::Commands::Definition::ExposeSelf.new(
       caller_location, env: env, format: format
@@ -270,7 +272,7 @@ module Qonfig::DSL # rubocop:disable Metrics/ModuleLength
   # @api public
   # @since 0.17.0
   def values_file(file_path, format: :dynamic, strict: false, expose: nil)
-    caller_location = caller(1, 1).first
+    caller_location = ::Kernel.caller(1, 1).first
 
     instance_commands << Qonfig::Commands::Instantiation::ValuesFile.new(
       file_path, caller_location, format: format, strict: strict, expose: expose