+++ lib/qonfig/imports/direct_key.rb
@@ -42,7 +42,7 @@ class Qonfig::Imports::DirectKey < Qonfig::Imports::Abstract
         "Setting with <#{key_matcher.scope_pattern}> key does not exist!"
       ) unless (imported_config.keys(all_variants: true).any? do |setting_key|
         key_matcher.match?(setting_key)
-      end)
+      end || key_matcher.generic?)
 
       imported_config.keys(all_variants: true).each do |setting_key|
         next unless key_matcher.match?(setting_key)
@@ -59,15 +59,16 @@ class Qonfig::Imports::DirectKey < Qonfig::Imports::Abstract
             define_method(access_method_name) do
               imported_config.slice_value(*setting_key_path_sequence)
             end
-
-            define_method("#{access_method_name}=") do |value|
-              imported_config
-            end
           else
             # NOTE: get setting object (concrete value or Qonfig::Settings object)
             define_method(access_method_name) do
               imported_config.dig(*setting_key_path_sequence)
             end
+          end
+
+          define_method("#{access_method_name}?") do
+            # NOTE: based on Qonfig::Settings#__define_option_predicate__ realization
+            !!imported_config[setting_key]
           end
 
           if accessor